refactor(v2): extract out into standalone components
