+++ v2/lib/theme/Docs/index.js
@@ -2,8 +2,11 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
 import Helmet from 'react-helmet';
-import styles from './styles.css';
+
 import Layout from '@theme/Layout'; // eslint-disable-line
+import DocsPaginator from '@theme/DocsPaginator'; // eslint-disable-line
+
+import styles from './styles.css';
 
 export default class Docs extends React.Component {
   renderSidebar(metadata, docsSidebars, docsMetadatas) {
@@ -62,22 +65,8 @@ export default class Docs extends React.Component {
           {version && <meta name="docsearch:version" content={version} />}
         </Helmet>
         <div>{this.renderSidebar(metadata, docsSidebars, docsMetadatas)}</div>
-        <div>
-          {metadata.previous &&
-            docsMetadatas[metadata.previous] && (
-              <Link to={docsMetadatas[metadata.previous].permalink}>
-                <span>← {metadata.previous_title}</span>
-              </Link>
-            )}
-          {' ⚫️ '}
-          {metadata.next &&
-            docsMetadatas[metadata.next] && (
-              <Link to={docsMetadatas[metadata.next].permalink}>
-                <span>{metadata.next_title} →</span>
-              </Link>
-            )}
-        </div>
         <div className={styles.mainContainer}>{this.props.children}</div>
+        <DocsPaginator docsMetadatas={docsMetadatas} metadata={metadata} />
       </Layout>
     );
   }
