+++ v2/lib/theme/Footer/styles.css
@@ -1,0 +1,39 @@
+.footer {
+  color: #777;
+  margin-top: 30px;
+  padding: 10px 15px;
+  height: 20px;
+  text-align: center;
+  border-top: 1px solid #e6e6e6;
+}
+
+.routeLinks {
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  position: absolute;
+  right: 0;
+  left: 0;
+}
+
+.routeLinks li {
+  display: inline;
+}
+
+.routeLinks li a {
+  color: inherit;
+  margin: 3px;
+  padding: 3px 7px;
+  text-decoration: none;
+  border: 1px solid transparent;
+  border-radius: 3px;
+}
+
+.routeLinks li a.selected,
+.routeLinks li a:hover {
+  border-color: rgba(175, 47, 47, 0.1);
+}
+
+.routeLinks li a.selected {
+  border-color: rgba(175, 47, 47, 0.2);
+}
