+++ v2/lib/theme/Layout/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
-import {Link} from 'react-router-dom';
-import styles from './styles.css';
+import Footer from '@theme/Footer'; // eslint-disable-line
 
 /* eslint-disable react/prefer-stateless-function */
 export default class Layout extends React.Component {
@@ -11,21 +10,14 @@ export default class Layout extends React.Component {
       docsMetadatas = {},
       location,
     } = this.props;
-    const docsFlatMetadatas = Object.values(docsMetadatas);
-    const routeLinks = [...pagesMetadatas, ...docsFlatMetadatas].map(
-      data =>
-        data.permalink !== location.pathname && (
-          <li key={data.permalink}>
-            <Link to={data.permalink}>{data.permalink}</Link>
-          </li>
-        ),
-    );
     return (
       <div>
         {children}
-        <div className={styles.footer}>
-          <ul className={styles.routeLinks}>{routeLinks}</ul>
-        </div>
+        <Footer
+          docsMetadatas={docsMetadatas}
+          location={location}
+          pagesMetadatas={pagesMetadatas}
+        />
       </div>
     );
   }
